more updates from testing finally fixed ship position inaccuracies when logout/login few more updates to effects system - works very well, but some minor problems still exist

+++ src/eve-server/inventory/InventoryDB.cpp
@@ -508,7 +508,7 @@ bool InventoryDB::SaveItem(uint32 itemID, const ItemData &data) {
         data.typeID,
         data.ownerID,
         data.locationID,
-        (uint32)data.flag,
+        (uint16)data.flag,
         (data.contraband?1:0),
         (data.singleton?1:0),
         data.quantity,
